refactor: modularize chat and add LLM service
